Consolidate os* files
