+++ ddprof-lib/src/main/cpp/buffers.h
@@ -8,7 +8,7 @@
 
 #include <arpa/inet.h>
 
-#include "os_dd.h"
+#include "os.h"
 
 const int BUFFER_SIZE = 1024;
 const int BUFFER_LIMIT = BUFFER_SIZE - 128;
