+++ ddprof-lib/src/main/cpp/codeCache.cpp
@@ -5,7 +5,7 @@
 
 #include "codeCache.h"
 #include "dwarf_dd.h"
-#include "os_dd.h"
+#include "os.h"
 
 #include <stdint.h>
 #include <stdlib.h>
