+++ ddprof-lib/src/main/cpp/context.cpp
@@ -16,7 +16,7 @@
 
 #include "context.h"
 #include "counters.h"
-#include "os_dd.h"
+#include "os.h"
 #include "thread.h"
 #include <cstring>
