+++ ddprof-lib/src/main/cpp/context.h
@@ -20,7 +20,7 @@
 #include "arch_dd.h"
 #include "arguments.h"
 #include "common.h"
-#include "os_dd.h"
+#include "os.h"
 #include "vmEntry.h"
 
 static const u32 DD_TAGS_CAPACITY = 10;
