+++ ddprof-lib/src/main/cpp/flightRecorder.cpp
@@ -17,7 +17,7 @@
 #include "jfrMetadata.h"
 #include "jniHelper.h"
 #include "jvm.h"
-#include "os_dd.h"
+#include "os.h"
 #include "profiler.h"
 #include "rustDemangler.h"
 #include "spinLock.h"
@@ -579,7 +579,7 @@ void Recording::switchChunk(int fd) {
   if (fd > -1) {
     // move the chunk to external file and reset the continuous recording file
     OS::copyFile(_fd, fd, 0, _chunk_start);
-    ddprof::OS::truncateFile(_fd);
+    OS::truncateFile(_fd);
     // need to reset the file offset here
     _chunk_start = 0;
     _base_id = 0;
