+++ ddprof-lib/src/main/cpp/j9Ext.cpp
@@ -16,7 +16,7 @@
  */
 
 #include "j9Ext.h"
-#include "os_dd.h"
+#include "os.h"
 #include <string.h>
 
 jvmtiEnv *J9Ext::_jvmti;
