+++ ddprof-lib/src/main/cpp/linearAllocator.cpp
@@ -17,7 +17,7 @@
 
 #include "linearAllocator.h"
 #include "counters.h"
-#include "os_dd.h"
+#include "os.h"
 #include "common.h"
 #include <stdio.h>
