+++ ddprof-lib/src/main/cpp/threadFilter.cpp
@@ -21,7 +21,7 @@
 
 #include "threadFilter.h"
 #include "arch_dd.h"
-#include "os_dd.h"
+#include "os.h"
 #include <cassert>
 #include <cstdlib>
 #include <cstdio>
