+++ ddprof-lib/src/main/cpp/threadInfo.h
@@ -1,5 +1,5 @@
 #include "mutex.h"
-#include "os_dd.h"
+#include "os.h"
 #include <functional>
 #include <map>
 #include <memory>
