+++ ddprof-lib/src/main/cpp/wallClock.h
@@ -8,7 +8,7 @@
 #define _WALLCLOCK_H
 
 #include "engine.h"
-#include "os_dd.h"
+#include "os.h"
 #include "profiler.h"
 #include "reservoirSampler.h"
 #include "thread.h"
