+++ ddprof-lib/src/test/cpp/ddprof_ut.cpp
@@ -5,7 +5,7 @@
     #include "context.h"
     #include "counters.h"
     #include "mutex.h"
-    #include "os_dd.h"
+    #include "os.h"
     #include "unwindStats.h"
     #include "threadFilter.h"
     #include "threadInfo.h"
