+++ ddprof-lib/src/test/cpp/safefetch_ut.cpp
@@ -4,7 +4,7 @@
 #include <signal.h>
 
 #include "safeAccess.h"
-#include "os_dd.h"
+#include "os.h"
 #include "../../main/cpp/gtest_crash_handler.h"
 
 // Test name for crash handler
@@ -31,13 +31,13 @@ void signal_handle_wrapper(int signo, siginfo_t* siginfo, void* context) {
 class SafeFetchTest : public ::testing::Test {
 protected:
     void SetUp() override {
-       orig_segvHandler = ddprof::OS::replaceSigsegvHandler(signal_handle_wrapper);
-       orig_busHandler = ddprof::OS::replaceSigbusHandler(signal_handle_wrapper);
+       orig_segvHandler = OS::replaceSigsegvHandler(signal_handle_wrapper);
+       orig_busHandler = OS::replaceSigbusHandler(signal_handle_wrapper);
     }
 
     void TearDown() override {
-       ddprof::OS::replaceSigsegvHandler(orig_segvHandler);
-       ddprof::OS::replaceSigbusHandler(orig_busHandler);
+       OS::replaceSigsegvHandler(orig_segvHandler);
+       OS::replaceSigbusHandler(orig_busHandler);
     }
 };
